Move setting initial animation order out of domReady.
