+++ assets/src/helpers.js
@@ -3,7 +3,7 @@
  */
 import uuid from 'uuid/v4';
 import classnames from 'classnames';
-import { each, every } from 'lodash';
+import { each, every, isEqual } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -18,6 +18,8 @@ import { getColorClassName, getColorObjectByAttributeValues, RichText } from '@w
 /**
  * Internal dependencies
  */
+// eslint-disable-next-line no-unused-vars
+import store from './stores/amp-story';
 import {
 	BlockNavigation,
 	EditorCarousel,
@@ -42,7 +44,19 @@ const {
 	getBlockRootClientId,
 	getBlockOrder,
 	getBlock,
+	getClientIdsWithDescendants,
 } = select( 'core/editor' );
+
+const {
+	addAnimation,
+	changeAnimationType,
+	changeAnimationDuration,
+	changeAnimationDelay,
+} = dispatch( 'amp/story' );
+
+const {
+	getAnimatedBlocks,
+} = select( 'amp/story' );
 
 const { updateBlockAttributes } = dispatch( 'core/editor' );
 
@@ -979,5 +993,29 @@ export const maybeSetTagName = ( clientId ) => {
 
 	if ( block.attributes.tagName !== tagName ) {
 		updateBlockAttributes( clientId, { tagName } );
+	}
+};
+
+/**
+ * Initializes the animations if it hasn't been done yet.
+ */
+export const maybeInitializeAnimations = () => {
+	const animations = getAnimatedBlocks();
+	if ( isEqual( {}, animations ) ) {
+		const allBlocks = getBlocksByClientId( getClientIdsWithDescendants() );
+		for ( const block of allBlocks ) {
+			const page = getBlockRootClientId( block.clientId );
+
+			if ( page ) {
+				const { ampAnimationType, ampAnimationAfter, ampAnimationDuration, ampAnimationDelay } = block.attributes;
+				const predecessor = allBlocks.find( ( b ) => b.attributes.anchor === ampAnimationAfter );
+
+				addAnimation( page, block.clientId, predecessor ? predecessor.clientId : undefined );
+
+				changeAnimationType( page, block.clientId, ampAnimationType );
+				changeAnimationDuration( page, block.clientId, ampAnimationDuration ? parseInt( ampAnimationDuration.replace( 'ms', '' ) ) : undefined );
+				changeAnimationDelay( page, block.clientId, ampAnimationDelay ? parseInt( ampAnimationDelay.replace( 'ms', '' ) ) : undefined );
+			}
+		}
 	}
 };
